Refactored test providers
