+++ core/test/java/game/FreedomLineTests.java
@@ -25,7 +25,7 @@ public class FreedomLineTests {
     private final Board<Stone> board = BoardUtils.parseBoardFromString(printedBoard, 8, 8);
 
     @ParameterizedTest
-    @MethodSource("providers.FreedomLineProviders#provideInitialPosition")
+    @MethodSource("game.providers.FreedomLineProviders#provideInitialPosition")
     void testFreedomLineCustomConstructor(int row, int column, Class<Exception> expectedException) {
         if (expectedException == null) {
             Assertions.assertDoesNotThrow(() -> new FreedomLine(board, Position.fromCoordinates(row, column)));
@@ -35,7 +35,7 @@ public class FreedomLineTests {
     }
 
     @ParameterizedTest
-    @MethodSource("providers.FreedomLineProviders#provideSetOfPositions")
+    @MethodSource("game.providers.FreedomLineProviders#provideSetOfPositions")
     void testAddMethod(SortedSet<Position> positions, Class<Exception> expectedException) {
         if (expectedException == null) {
             Assertions.assertDoesNotThrow(() -> {
